Modularize runtime/SDL compilers and clear SonarCloud new-code smells
Drive all 180 SonarCloud new-code findings on the refactored files to zero, behavior-preserving. Split the three god-files under 500 lines each: compiler.py (1431) into compiler + 5 compiler_* modules, validator.py (1342) into a slim orchestrator + 6 validator_* mixin modules, planner.py (958) into planner + 4 planner_* modules; tidy parser.py. Add docstrings and specific type hints throughout, reduce per-function cognitive/cyclomatic complexity (all changed code now <=9 McCabe), cut compile_runtime_model to 98 lines, and fix assorted smells (too-many-returns/params, static method, dict.fromkeys, merged ifs, trailing comments, single Array.push in workbench.ts). Public APIs unchanged; full pytest (1776) and vitest suites stay green.
