+++ src/aptl/core/sdl/parser.py
@@ -6,8 +6,11 @@ Provides ``parse_sdl()`` as the primary entry point. Handles:
 - Shorthand expansion (``source: "pkg"`` → ``{name: "pkg", version: "*"}``)
 """
 
+from __future__ import annotations
+
+from collections.abc import Hashable
 from pathlib import Path
-from typing import Any
+from typing import TypeAlias
 
 import yaml
 from pydantic import ValidationError
@@ -16,6 +19,18 @@ from aptl.core.sdl._errors import SDLParseError, SDLValidationError
 from aptl.core.sdl._base import is_variable_ref
 from aptl.core.sdl.scenario import Scenario
 from aptl.core.sdl.validator import SemanticValidator
+
+
+# A value as produced by ``yaml.safe_load``: scalars, lists and mappings
+# nested to arbitrary depth. Mapping keys may be non-string scalars, so the
+# alias is recursive on both keys and values.
+YAMLValue: TypeAlias = (
+    "None | bool | int | float | str "
+    "| list[YAMLValue] | dict[Hashable, YAMLValue]"
+)
+
+# A normalized SDL mapping: top-level YAML document after key normalization.
+YAMLMapping: TypeAlias = "dict[Hashable, YAMLValue]"
 
 
 # Top-level sections that are HashMaps of user-defined identifiers.
@@ -32,18 +47,26 @@ _HASHMAP_SECTIONS = frozenset({
 
 # Fields within struct models that are also HashMaps of user-defined keys.
 _NESTED_HASHMAP_FIELDS = frozenset({
-    "features",            # VM.features (dict[str, str])
-    "conditions",          # VM.conditions (dict[str, str])
-    "injects",             # VM.injects (dict[str, str])
-    "roles",               # Node.roles (dict[str, Role])
-    "facts",               # Entity.facts (dict[str, str])
-    "entities",            # Entity.entities (dict[str, Entity])
-    "events",              # Script.events (dict[str, int])
-    "steps",               # Workflow.steps (dict[str, WorkflowStep])
+    # VM.features (dict[str, str])
+    "features",
+    # VM.conditions (dict[str, str])
+    "conditions",
+    # VM.injects (dict[str, str])
+    "injects",
+    # Node.roles (dict[str, Role])
+    "roles",
+    # Entity.facts (dict[str, str])
+    "facts",
+    # Entity.entities (dict[str, Entity])
+    "entities",
+    # Script.events (dict[str, int])
+    "events",
+    # Workflow.steps (dict[str, WorkflowStep])
+    "steps",
 })
 
 
-def _child_is_hashmap_field(key: str, value: Any) -> bool:
+def _child_is_hashmap_field(key: str, value: YAMLValue) -> bool:
     """Return whether the children of ``key`` are user-defined hashmap keys."""
     if key in _HASHMAP_SECTIONS or key in _NESTED_HASHMAP_FIELDS:
         return True
@@ -51,14 +74,14 @@ def _child_is_hashmap_field(key: str, value: Any) -> bool:
     return key == "properties" and isinstance(value, list)
 
 
-def _normalize_field_key(k: Any) -> Any:
+def _normalize_field_key(k: Hashable) -> Hashable:
     """Normalize a Pydantic field key: lowercase + hyphens to underscores."""
     if isinstance(k, str):
         return k.lower().replace("-", "_")
     return k
 
 
-def _normalize_keys(data: Any, is_hashmap: bool = False) -> Any:
+def _normalize_keys(data: YAMLValue, is_hashmap: bool = False) -> YAMLValue:
     """Normalize dict keys for Pydantic field matching.
 
     Pydantic struct field keys are lowercased with hyphens converted to
@@ -87,52 +110,82 @@ def _normalize_keys(data: Any, is_hashmap: bool = False) -> Any:
     return data
 
 
+def _reject_mapping_key_if_variable(
+    key: Hashable,
+    *,
+    path: str,
+    is_hashmap: bool,
+) -> None:
+    """Raise if ``key`` is a ``${var}`` placeholder in a symbol position."""
+    if not (is_hashmap and is_variable_ref(key)):
+        return
+    key_path = f"{path}.{key}" if path else str(key)
+    raise SDLParseError(
+        "Variable placeholders are not allowed in "
+        f"user-defined mapping keys: '{key_path}'"
+    )
+
+
+def _reject_in_mapping(
+    data: dict[Hashable, YAMLValue],
+    *,
+    path: str,
+    is_hashmap: bool,
+) -> None:
+    """Recurse over a mapping, rejecting variable placeholders in its keys."""
+    for k, v in data.items():
+        _reject_mapping_key_if_variable(k, path=path, is_hashmap=is_hashmap)
+
+        child_key = k if isinstance(k, str) else str(k)
+        child_path = f"{path}.{child_key}" if path else child_key
+        child_is_hashmap = (
+            False if is_hashmap else _child_is_hashmap_field(child_key, v)
+        )
+        _reject_variable_mapping_keys(
+            v,
+            path=child_path,
+            is_hashmap=child_is_hashmap,
+        )
+
+
+def _reject_in_sequence(
+    data: list[YAMLValue],
+    *,
+    path: str,
+    is_hashmap: bool,
+) -> None:
+    """Recurse over a sequence, propagating the hashmap flag to items."""
+    for index, item in enumerate(data):
+        _reject_variable_mapping_keys(
+            item,
+            path=f"{path}[{index}]",
+            is_hashmap=is_hashmap,
+        )
+
+
 def _reject_variable_mapping_keys(
-    data: Any,
+    data: YAMLValue,
     *,
     path: str = "",
     is_hashmap: bool = False,
 ) -> None:
     """Reject ``${var}`` placeholders in symbol-defining mapping keys."""
     if isinstance(data, dict):
-        for k, v in data.items():
-            if is_hashmap and is_variable_ref(k):
-                key_path = f"{path}.{k}" if path else str(k)
-                raise SDLParseError(
-                    "Variable placeholders are not allowed in "
-                    f"user-defined mapping keys: '{key_path}'"
-                )
-
-            child_key = k if isinstance(k, str) else str(k)
-            child_path = f"{path}.{child_key}" if path else child_key
-            child_is_hashmap = (
-                False if is_hashmap else _child_is_hashmap_field(child_key, v)
-            )
-            _reject_variable_mapping_keys(
-                v,
-                path=child_path,
-                is_hashmap=child_is_hashmap,
-            )
-        return
-
-    if isinstance(data, list):
-        for index, item in enumerate(data):
-            child_path = f"{path}[{index}]"
-            _reject_variable_mapping_keys(
-                item,
-                path=child_path,
-                is_hashmap=is_hashmap,
-            )
+        _reject_in_mapping(data, path=path, is_hashmap=is_hashmap)
+    elif isinstance(data, list):
+        _reject_in_sequence(data, path=path, is_hashmap=is_hashmap)
 
 
-def _expand_source(value: Any) -> Any:
+def _expand_source(value: YAMLValue) -> YAMLValue:
     """Expand shorthand source: 'pkg-name' → {name: 'pkg-name', version: '*'}."""
     if isinstance(value, str):
         return {"name": value, "version": "*"}
     return value
 
 
-def _expand_infrastructure(infra: dict[str, Any]) -> dict[str, Any]:
+def _expand_infrastructure(
+    infra: dict[Hashable, YAMLValue],
+) -> dict[Hashable, YAMLValue]:
     """Expand infrastructure shorthand: {node: 3} → {node: {count: 3}}."""
     result = {}
     for name, value in infra.items():
@@ -143,7 +196,9 @@ def _expand_infrastructure(infra: dict[str, Any]) -> dict[str, Any]:
     return result
 
 
-def _expand_roles(roles: dict[str, Any]) -> dict[str, Any]:
+def _expand_roles(
+    roles: dict[Hashable, YAMLValue],
+) -> dict[Hashable, YAMLValue]:
     """Expand role shorthand: {admin: 'username'} → {admin: {username: 'username'}}."""
     result = {}
     for name, value in roles.items():
@@ -154,7 +209,7 @@ def _expand_roles(roles: dict[str, Any]) -> dict[str, Any]:
     return result
 
 
-def _expand_min_score(value: Any) -> Any:
+def _expand_min_score(value: YAMLValue) -> YAMLValue:
     """Expand min-score shorthand: 50 → {percentage: 50}."""
     if isinstance(value, int) or is_variable_ref(value):
         return {"percentage": value}
@@ -166,11 +221,11 @@ _SOURCE_SKIP_SECTIONS = frozenset({"relationships", "agents"})
 
 
 def _expand_sources_scoped(
-    obj: Any,
+    obj: YAMLValue,
     *,
     is_hashmap: bool = False,
     skip: bool = False,
-) -> Any:
+) -> YAMLValue:
     """Recursively expand ``source`` shorthand outside of skipped sections."""
     if isinstance(obj, dict):
         result = {}
@@ -207,7 +262,7 @@ def _expand_sources_scoped(
     return obj
 
 
-def _expand_node_shorthands(nodes: dict[str, Any]) -> None:
+def _expand_node_shorthands(nodes: dict[Hashable, YAMLValue]) -> None:
     """Expand roles and feature/condition/inject list shorthands within nodes."""
     for node_data in nodes.values():
         if isinstance(node_data, dict):
@@ -216,12 +271,10 @@ def _expand_node_shorthands(nodes: dict[str, Any]) -> None:
             # G6: features/conditions/injects as list -> dict with empty role
             for field in ("features", "conditions", "injects"):
                 if field in node_data and isinstance(node_data[field], list):
-                    node_data[field] = {
-                        name: "" for name in node_data[field]
-                    }
+                    node_data[field] = dict.fromkeys(node_data[field], "")
 
 
-def _expand_evaluation_shorthands(evaluations: dict[str, Any]) -> None:
+def _expand_evaluation_shorthands(evaluations: dict[Hashable, YAMLValue]) -> None:
     """Expand min_score shorthand in evaluations."""
     for eval_data in evaluations.values():
         if isinstance(eval_data, dict) and "min_score" in eval_data:
@@ -230,7 +283,7 @@ def _expand_evaluation_shorthands(evaluations: dict[str, Any]) -> None:
             )
 
 
-def _expand_shorthands(data: dict[str, Any]) -> dict[str, Any]:
+def _expand_shorthands(data: YAMLMapping) -> YAMLMapping:
     """Apply all shorthand expansions to normalized data."""
     data = _expand_sources_scoped(data)
 
@@ -318,7 +371,7 @@ def parse_sdl(
     return scenario
 
 
-def parse_sdl_file(path: Path, **kwargs: Any) -> Scenario:
+def parse_sdl_file(path: Path, **kwargs: bool) -> Scenario:
     """Parse an SDL YAML file into a validated Scenario.
 
     Convenience wrapper around ``parse_sdl()`` that reads from a file.
